move random number bound to IRandomService api

+++ whirl/engines/matrix/world/global/random.cpp
@@ -6,12 +6,12 @@ namespace whirl::matrix {
 
 // size_t GlobalRandomNumber() defined in global.cpp
 
-size_t GlobalRandomNumber(size_t bound) {
+uint64_t GlobalRandomNumber(uint64_t bound) {
   WHEELS_VERIFY(bound > 0, "bound = 0");
   return GlobalRandomNumber() % bound;
 }
 
-size_t GlobalRandomNumber(size_t lo, size_t hi) {
+uint64_t GlobalRandomNumber(uint64_t lo, uint64_t hi) {
   WHEELS_VERIFY(lo <= hi, "Invalid range");
   return lo + GlobalRandomNumber(hi - lo);
 }
